requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
flask[async]>=3.0.0
pydantic>=2.5.2
pytest>=7.4.3
pytest-asyncio>=0.23.2
flasgger>=0.9.5
gunicorn>=21.2.0
uvicorn>=0.25.0
redis>=5.0.0
celery>=5.3.0
tenacity>=8.2.0
flask-cors>=4.0.0
